[Silver III] Title: 주유소, Time: 448 ms, Memory: 36684 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -8,17 +8,17 @@ public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         int n = Integer.parseInt(br.readLine());
-        int[] distance = new int[n - 1];
+        long[] distance = new long[n - 1];
         StringTokenizer st = new StringTokenizer(br.readLine());
         for (int i = 0; i < n - 1; i++) {
             distance[i] = Integer.parseInt(st.nextToken());
         }
         st = new StringTokenizer(br.readLine());
-        int[] liter = new int[n];
+        long[] liter = new long[n];
         for (int i = 0; i < n; i++) {
             liter[i] = Integer.parseInt(st.nextToken());
         }
-        int min = liter[0];
+        long min = liter[0];
         long sum = 0;
         for (int i = 0; i < n - 1; i++) {
             if (liter[i] < min) {
